pyproject.toml
[tool.poetry]
name = "expensetracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.34.0"
streamlit-option-menu = "^0.3.12"
plotly = "^5.22.0"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"
streamlit-authenticator = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
